internalize tera engine because we already dep on tera
