+++ src/controller/views/mod.rs
@@ -1,3 +1,4 @@
+pub mod engines;
 use axum::{async_trait, extract::FromRequestParts, http::request::Parts, Extension};
 use serde::Serialize;
